backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0",
    "flask-login>=0.6",
    "flask-restx>=1.3",
    "flask-sqlalchemy>=3.1",
    "flask-migrate>=4.0",
    "psycopg2-binary>=2.9",
    "werkzeug>=3.0",
    "anthropic>=0.52",
]
